restructure library
