+++ src/api.js
@@ -1,160 +1,4 @@
-import { NativeModules, TouchableWithoutFeedbackBase } from "react-native";
 
-const RNFastStorage = NativeModules.RNFastStorage;
-
-
-const specials = '_.][?)(&$/{}';
-const lowercase = 'abcdefghijklmnopqrstuvwxyz';
-const uppercase = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
-const numbers = '0123456789';
-
-const all = specials + lowercase + uppercase + numbers;
-
-export default function generatePassword() {
-  let password = '';
-
-  password += pick(password, specials, 1, 3);
-  password += pick(password, lowercase, 1, 3);
-  password += pick(password, uppercase, 1, 3);
-  password += pick(password, all, 10);
-
-  return shuffle(password);
-}
-
-function pick(exclusions, string, min, max) {
-  var n, chars = '';
-
-  if (max === undefined) {
-    n = min;
-  } else {
-    n = min + Math.floor(Math.random() * (max - min + 1));
-  }
-
-  var i = 0;
-  while (i < n) {
-    const character = string.charAt(Math.floor(Math.random() * string.length));
-    if (exclusions.indexOf(character) < 0 && chars.indexOf(character) < 0) {
-      chars += character;
-      i++;
-    }
-  }
-
-  return chars;
-}
-
-const stringToHex = (input) => {
-  let str = ""
-  for (const char of input) {
-    str += char.charCodeAt(0).toString(16)
-  }
-  return str
-}
-
-function shuffle(string) {
-  var array = string.split('');
-  var tmp, current, top = array.length;
-
-  if (top) while (--top) {
-    current = Math.floor(Math.random() * (top + 1));
-    tmp = array[current];
-    array[current] = array[top];
-    array[top] = tmp;
-  }
-
-  return array.join('');
-}
-
-
-
-export class loader {
-  constructor() {
-      this.instanceID = null;
-      this.initWithEncryption = false;
-      this.secureKeyStorage = false;
-      this.accessibleModeForSecureKey = null;
-      this.mmkvDirPath = '';
-      this.initWithCustomDirPath = false;
-      this.processingMode = '';
-      this.defaultAlias = 'com.ammarahmed.MMKV';
-      this.alias = null;
-      this.key = null;
-  }
-
-default() {
-  RNFastStorage.setupLibrary();
-}
-
-withInstanceID(id) {
-  this.instanceID = id;
-
-  return this;
-}
-
-withEncryption() {
-  this.initWithEncryption = true;
-  this.key = this.generateKey();
-  this.alias = this.defaultAlias;
-  this.secureKeyStorage = true;
-  return this;
-}
-
-withCustomKey(key) {
-  this.key = key;
-  this.secureKeyStorage = false;
-  return this;
-}
-
-
-withSecureKeyStorage(alias = this.defaultAlias) {
-  this.secureKeyStorage = true;
-  this.alias = stringToHex(alias);
-
-  return this;
-}
-
-withCustomDirPath(path) {
-  this.initWithCustomDirPath = true;
-  this.mmkvDirPath = path;
-
-  return this;
-}
-
-setProcessingMode(mode) {
-  this.processingMode = mode;
-
-  return this;
-}
-
-initialize() {
- console.log(this);
-}
-
-generateKey() {
-  this.key = generatePassword();
-
-  return this;
-}
-
-encrypt(key) {
- RNFastStorage.encrypt(key);
-}
-
-decrypt() {
-  RNFastStorage.decrypt();
-}
-
-changeEncryptionKey(key) {
-  RNFastStorage.encrypt(key);
-}
-
-}
-
-
-
-
-
-
-if (RNFastStorage.setupLibrary) RNFastStorage.setupLibrary();
 
 /**
  * Set a string value to storage for a given key.
