+++ src/helpers.js
@@ -1,0 +1,8 @@
+
+const stringToHex = (input) => {
+  let str = ""
+  for (const char of input) {
+    str += char.charCodeAt(0).toString(16)
+  }
+  return str
+}
